+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 - mobile: show developer menu is now saved on disk and remebered between start of Subsurface
 - tests: add qml test harness
+- Dive media: experimental support for metadata extraction from AVI and WMV files
 - Dive media: sort thumbnails by timestamp
 - Dive media: don't recalculate all pictures on drag & drop
 - Profile: immediately update thumbnail positions on deletion